Move search results into a reusable component
