+++ app/components/Search/SearchPageResults.css
@@ -1,5 +1,3 @@
-@import 'app/styles/variables.css';
-
 .searchResult {
   padding-left: 20px;
   border-left: 4px solid transparent;
